+++ app/javascript/dashboard/helper/scriptGenerator.js
@@ -1,20 +1,3 @@
-export const createWebsiteWidgetScript = websiteToken => `
-<script>
-(function(d,t) {
-  var BASE_URL = '${window.location.origin}';
-  var g=d.createElement(t),s=d.getElementsByTagName(t)[0];
-  g.src= BASE_URL + "/packs/js/sdk.js";
-  s.parentNode.insertBefore(g,s);
-  g.onload=function(){
-    window.chatwootSDK.run({
-      websiteToken: '${websiteToken}',
-      baseUrl: BASE_URL
-    })
-  }
-})(document,"script");
-</script>
-`;
-
 export const createMessengerScript = pageId => `
 <script>
   window.fbAsyncInit = function() {
