+++ app/javascript/dashboard/routes/dashboard/settings/inbox/Settings.vue
@@ -22,7 +22,7 @@
           :sub-title="$t('INBOX_MGMT.SETTINGS_POPUP.MESSENGER_SUB_HEAD')"
         >
         </settings-form-header>
-        <woot-code :script="webWidgetScript"></woot-code>
+        <woot-code :script="inbox.web_widget_script"></woot-code>
       </div>
       <div class="settings--content">
         <settings-form-header
@@ -70,8 +70,12 @@
       >
       </settings-form-header>
       <select v-model="autoAssignment">
-        <option value="true">{{ $t('INBOX_MGMT.EDIT.AUTO_ASSIGNMENT.ENABLED') }}</option>
-        <option value="false">{{ $t('INBOX_MGMT.EDIT.AUTO_ASSIGNMENT.DISABLED') }}</option>
+        <option value="true">
+          {{ $t('INBOX_MGMT.EDIT.AUTO_ASSIGNMENT.ENABLED') }}
+        </option>
+        <option value="false">
+          {{ $t('INBOX_MGMT.EDIT.AUTO_ASSIGNMENT.DISABLED') }}
+        </option>
       </select>
     </div>
   </div>
@@ -81,10 +85,7 @@
 /* eslint no-console: 0 */
 /* global bus */
 import { mapGetters } from 'vuex';
-import {
-  createWebsiteWidgetScript,
-  createMessengerScript,
-} from 'dashboard/helper/scriptGenerator';
+import { createMessengerScript } from 'dashboard/helper/scriptGenerator';
 import { Compact } from 'vue-color';
 import SettingsFormHeader from '../../../../components/SettingsFormHeader.vue';
 
@@ -111,9 +112,6 @@ export default {
     },
     inbox() {
       return this.$store.getters['inboxes/getInbox'](this.currentInboxId);
-    },
-    webWidgetScript() {
-      return createWebsiteWidgetScript(this.inbox.website_token);
     },
     messengerScript() {
       return createMessengerScript(this.inbox.page_id);
@@ -196,9 +194,13 @@ export default {
             enable_auto_assignment: this.autoAssignment,
           },
         });
-        this.showAlert(this.$t('INBOX_MGMT.EDIT.API.AUTO_ASSIGNMENT_SUCCESS_MESSAGE'));
+        this.showAlert(
+          this.$t('INBOX_MGMT.EDIT.API.AUTO_ASSIGNMENT_SUCCESS_MESSAGE')
+        );
       } catch (error) {
-        this.showAlert(this.$t('INBOX_MGMT.EDIT.API.AUTO_ASSIGNMENT_SUCCESS_MESSAGE'));
+        this.showAlert(
+          this.$t('INBOX_MGMT.EDIT.API.AUTO_ASSIGNMENT_SUCCESS_MESSAGE')
+        );
       }
     },
     getWidgetColor() {
